rearrange package a bit
