+++ packages/neutron-sdk/src/interchain_txs/mod.rs
@@ -1,3 +1,4 @@
 pub mod helpers;
 
+pub mod v045;
 pub mod v047;
